requirements.txt
django
django-widget-tweaks
django-jazzmin
django-role-permissions